modify the behaviour of _log_info
