+++ src/pytorch_lightning/callbacks/early_stopping.py
@@ -256,17 +256,3 @@ class EarlyStopping(Callback):
         else:
             msg = f"Metric {self.monitor} improved. New best score: {current:.3f}"
         return msg
-
-    @staticmethod
-    def _log_info(trainer: Optional["pl.Trainer"], message: str, log_rank_zero_only: bool) -> None:
-        if trainer:
-            # ignore logging in non-zero ranks if log_rank_zero_only flag is enabled
-            if log_rank_zero_only and trainer.global_rank != 0:
-                return
-            # if world size is more than one then specify the rank of the process being logged
-            if trainer.world_size > 1:
-                log.info(f"[rank: {trainer.global_rank}] {message}")
-                return
-
-        # if above conditions don't meet and we have to log
-        log.info(message)
